refactor(graph): the last three query families — queries.ts is a facade (#1838)
`queries.ts` is 61 lines and carries no logic: six families under
`graph/queries/`, and a file whose only job is that `graph/index` and the ~40
`import … from './queries'` call sites don't have to know which family a
function belongs to.

  1,294 → 377 (tags, sources, links) → 61 (notes, references, sparql)

The three that moved:

- `notes.ts` — what the store knows about a note AS a note: aliases both
  directions, the project's frontmatter vocabulary, its IRI, its headings.
- `references.ts` — "what points at this?", where the thing is a source, an
  excerpt or a heading anchor. Distinct from `links.ts`, which answers the same
  question for note-to-note edges. Also the two whole-project note listings the
  refactor and glossary paths walk.
- `sparql.ts` — the engine's own surface: prefix injection, the completion
  schema, `queryGraph`. The odd one out by design: every other family answers a
  specific question about notes; this one runs whatever the user asked.

Checked in both directions before moving, as with the earlier three: no family
calls into another, and each private helper is used only by its own family.
`findNotesLinkingToAnchorImpl` stays reachable through the facade, which is how
`indexers.ts` gets it — the one edge from the write side into the read side,
still one-way.

I argued in #1874 that the remaining ~340 lines were one concern and not worth
splitting. Having done it: the SPARQL third is clearly its own thing, and
separating "what does this note look like" from "what points at it" reads
better than I expected. The facade's header now says what makes each family a
family, which is the part that was actually missing.

Whole suite passes with no test edits — the same proof the earlier extractions
carried.
